changed ExpenseItem amount to Double instead of Float

+++ MyFinancesApp/MyFinances/Expenses API/ExpenseItemsMapper.swift
@@ -37,7 +37,7 @@ extension Root: Decodable {
         for key in container.allKeys {
             let productContainer = try container.nestedContainer(keyedBy: ExpenseKey.self, forKey: key)
             let title = try productContainer.decode(String.self, forKey: .title)
-            let amount = try productContainer.decode(Float.self, forKey: .amount)
+            let amount = try productContainer.decode(Double.self, forKey: .amount)
             let created_at = try productContainer.decode(Date.self, forKey: .created_at)
             let expense = ExpenseItem(id: UUID(uuidString: key.stringValue)!, title: title, amount: amount, createdAt: created_at)
             expenses.append(expense)
